Bring Cache Tags and Cache Events back from 5.0
As Cache Tags are a big help for me and I only found documentation in the 5.0 docs, I brought these back in to the 5.1 documentation. Doing so I noticed the Cache Events were also missing, so I brought those back as well. 
It seems they are still working, so I hope this was an OK move...