Add the missing critical info about how to set up the server
This chapter missed the most critical info about how to get up and run the newly created project quickly and easily, this breaks the entire flow of the learning process. The sections and chapters followed are all not mentioning a quick and easy way about how to set up this thing besides the complex Homestand and Valet. How can we not added this already? 

Come on, get the developers excited by adding this to the doc.