update slot documentation
passing strings through the slots is incredibly easy and well documented, but I ran into an issue when trying to pass PHP constructs through, like arrays or objects.  I was able to pass it as the second argument to the slot, and it worked fine.  I'm curious if this is the intended operation, and if so I'd like to add some documentation on it.