Update v5.4 Upgrade Guide
Currently the Upgrade guide references a change in the API to
relations where the name of the `getForeignKey` method has changed
to `getForeignKeyName`  In reviewing the code and API docs this
doesn't seem to actually be true.

Example (https://laravel.com/api/5.4/Illuminate/Database/Eloquent/Relations/BelongsTo.html)

This commit removes the seemingly false reference.